requirements.txt
Flask==2.2.2
Jinja2==3.1.2
markupsafe==2.1.1
itsdangerous==2.1.2
werkzeug==2.2.2
psycopg2==2.9.5
flask-sqlalchemy==3.0.2
Flask-Migrate==4.0.2
psycopg2-binary==2.9.5